arduino-dashboard/requirements.txt
streamlit==1.26.0
pyserial==3.5
pandas==2.1.2
altair==5.2.0
python-dotenv==1.0.0
